create single build definition, to (eventually!) allow more efficient rebuilds
