+++ .jshintrc
@@ -12,6 +12,8 @@
 	"-W030": false,
 	"esnext": true,
 	"globals": {
+		"process": true,
+		"__dirname": true,
 		"clearInterval": true,
 		"clearTimeout": true,
 		"define": true,
